+++ dyninstAPI/src/BPatch/BPatch_binaryEdit.C
