+++ dyninstAPI/src/BPatch/BPatch_collections.C
