+++ dyninstAPI/src/BPatch/BPatch_collections.h
