+++ dyninstAPI/src/BPatch/BPatch_edge.C
