+++ dyninstAPI/src/BPatch/BPatch_frame.C
