+++ dyninstAPI/src/BPatch/BPatch_function.C
