+++ dyninstAPI/src/BPatch/BPatch_image.C
