+++ dyninstAPI/src/BPatch/BPatch_instruction.C
