+++ dyninstAPI/src/BPatch/BPatch_libInfo.h
