+++ dyninstAPI/src/BPatch/BPatch_memoryAccess.C
