+++ dyninstAPI/src/BPatch/BPatch_memoryAccessAdapter.C
