+++ dyninstAPI/src/BPatch/BPatch_memoryAccessAdapter.h
