+++ dyninstAPI/src/BPatch/BPatch_module.C
