+++ dyninstAPI/src/BPatch/BPatch_object.C
