+++ dyninstAPI/src/BPatch/BPatch_private.h
