+++ dyninstAPI/src/BPatch/BPatch_snippet.C
