+++ dyninstAPI/src/BPatch/BPatch_thread.C
