+++ dyninstAPI/src/BPatch/BPatch_type.C
